Suck_Company/web_app/modelsserializers.py
from .models import Company,Category
from rest_framework import permissions, routers, serializers, viewsets

class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = '__all__'

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'
