requirements.txt
django==1.11.1
django-filter==1.1.0
djangorestframework==3.7.1
pygments==2.2.0
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
